+++ src/main.c
@@ -105,6 +105,7 @@ accountSender_t global_account_sender;
 #define INS_UPDATE_LEVEL1_KEYS 0x29
 #define INS_UPDATE_LEVEL2_KEYS_ROOT 0x2A
 #define INS_UPDATE_LEVEL2_KEYS_LEVEL1 0x2B
+#define INS_ADD_ANONYMITY_REVOKER 0x2C
 #define INS_ADD_IDENTITY_PROVIDER 0x2D
 
 #define INS_SIGN_UPDATE_CREDENTIAL      0x31
@@ -250,6 +251,9 @@ static void concordium_main(void) {
                         break;
                     case INS_ADD_IDENTITY_PROVIDER:
                         handleSignAddIdentityProvider(cdata, p1, lc, &flags, isInitialCall);
+                        break;
+                    case INS_ADD_ANONYMITY_REVOKER:
+                        handleSignAddAnonymityRevoker(cdata, p1, lc, &flags, isInitialCall);
                         break;
                     default:
                         THROW(ERROR_INVALID_INSTRUCTION);